backend/pyproject.toml
[build-system]
requires = ["setuptools"]

[project]
name = "neuroagent"
authors = [
    {name = "Blue Brain Project, EPFL"},
]
description = "Search agent for neuroscience."
readme = "README.md"
requires-python = ">=3.11"
dynamic = ["version"]
dependencies = [
    "boto3",
    "celery>=5.5.3",
    "datamodel-code-generator==0.49.0",
    "httpx>=0.28.1",
    "openai>=2.6.1",
    "pydantic>=2.11",
    "pydantic-settings",
    "python-dotenv",
    "redis",
]

[project.optional-dependencies]
app = [
    "alembic>=1.17.2",
    "asgi-correlation-id>=4.3.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.122.0",
    "mcp>=1.22.0",
    "obp-accounting-sdk>=0.4.2",
    "psycopg2>=2.9.11",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn>=0.38.0",
]
dev = [
    "bandit[toml]",
    "colorama>=0.4.6",
    "datamodel-code-generator>=0.40.0",
    "deepeval>=3.6.6",
    "mypy==1.15.0",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-httpx",
    "responses",
    "ruff>=0.14.8",
    "types-requests",
]
tasks = [
    "bluepysnap>=3.0.2",
    "duckdb>=1.4.2",
    "pandas>=2.3.3",
]

[project.scripts]

[tool.setuptools.dynamic]
version = {attr = "neuroagent.__version__"}

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.setuptools.package-data]
neuroagent = ["rules/*.mdc", "mcp.json"]

[tool.bandit]
exclude_dirs = ["tests"]

[tool.black]
line-length = 88
preview = true

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = ["src/neuroagent/tools/autogenerated_types"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "D", "RUF100"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D"]
"alembic/*" = ["D", "E"]

[tool.mypy]
mypy_path = "src"
python_version = "3.11"
ignore_missing_imports = true
no_implicit_optional = true
check_untyped_defs = true
strict_equality = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
exclude = 'src/neuroagent/tools/autogenerated_types/.*'
plugins = [
  "pydantic.mypy", "sqlalchemy.ext.mypy.plugin"
]

[tool.pydantic-mypy]
init_typed = true

[[tool.mypy.overrides]]
module = "neuroagent.tools.*"
disable_error_code = ["override"]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
]
filterwarnings = [
    "error",
    "ignore:Use get_feature_values:DeprecationWarning",
    "ignore:Mean of empty slice:RuntimeWarning",
    "ignore:Degrees of freedom:RuntimeWarning",
    "ignore:Default value *:pydantic.json_schema.PydanticJsonSchemaWarning",
    "ignore:Exception ignored in:pytest.PytestUnraisableExceptionWarning",
    "ignore:The configuration option 'asyncio_default_fixture_loop_scope' is unset.",
    "ignore:datetime.*is deprecated and scheduled for removal in a future version.:DeprecationWarning",
    "ignore:Support for class-based:pydantic.warnings.PydanticDeprecatedSince20",
]

addopts = "--cov=src/ -v --cov-report=term-missing --durations=20 --no-cov-on-fail"
